pyproject.toml
[project]
name = "aws-analytics-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-msk-iam-sasl-signer-python>=1.0.2",
    "boto3>=1.39.3",
    "kafka-python>=2.2.15",
    "opensearch-py>=3.0.0",
    "python-dotenv>=1.1.1",
    "requests-aws4auth>=1.3.1",
]
